+++ netstrip/gui/views/__init__.py
@@ -3,3 +3,11 @@ from .blocklists import BlocklistView
 from .logs import LogView
 from .settings import SettingsView
 from .connections import ConnectionsView
+
+__all__ = [
+    "AppRulesView",
+    "BlocklistView",
+    "LogView",
+    "SettingsView",
+    "ConnectionsView",
+]